+++ test-suite/src/base_factory/tests.rs
@@ -1,1 +1,2 @@
 mod integration_tests;
+mod sudo_tests;
