+++ test-suite/src/common_setup/templates.rs
@@ -7,6 +7,10 @@ use crate::common_setup::{
     setup_minter::vending_minter::setup::{configure_minter, vending_minter_code_ids},
 };
 
+use super::msg::{Accounts, MinterTemplateResponseCodeIds};
+use super::setup_minter::base_minter::setup::base_minter_sg721_collection_code_ids;
+use super::setup_minter::common::constants::{MINT_PRICE, MIN_MINT_PRICE};
+use super::setup_minter::common::minter_params::minter_params_all;
 use crate::common_setup::setup_minter::base_minter::setup::base_minter_sg721_nt_code_ids;
 use crate::common_setup::setup_minter::base_minter::setup::configure_base_minter;
 use crate::common_setup::setup_minter::common::constants::{
@@ -17,21 +21,14 @@ use crate::common_setup::setup_minter::open_edition_minter::mock_params::{
     mock_create_minter, mock_params_proper,
 };
 use crate::common_setup::setup_minter::open_edition_minter::setup::open_edition_minter_code_ids;
-use cosmwasm_std::{coin, coins, Addr, Coin, Timestamp, Uint128};
-use cw_multi_test::{AppResponse, BankSudo, Contract, Executor, SudoMsg};
+use crate::common_setup::setup_minter::vending_minter::setup::vending_minter_updatable_code_ids;
+use cosmwasm_std::{coin, coins, Coin, Timestamp, Uint128};
+use cw_multi_test::{AppResponse, Contract, Executor};
 use open_edition_factory::types::{NftData, NftMetadataType};
 use sg2::msg::Sg2ExecuteMsg;
 use sg2::tests::mock_collection_params_1;
 use sg_multi_test::StargazeApp;
 use sg_std::{StargazeMsgWrapper, GENESIS_MINT_START_TIME, NATIVE_DENOM};
-
-use super::msg::Accounts;
-use super::setup_minter::base_minter::setup::base_minter_sg721_collection_code_ids;
-use super::setup_minter::common::constants::{MINT_PRICE, MIN_MINT_PRICE};
-use super::setup_minter::common::minter_params::minter_params_all;
-use super::setup_minter::vending_minter::setup::{
-    sudo_update_params, vending_minter_updatable_code_ids,
-};
 
 pub fn vending_minter_template(num_tokens: u32) -> MinterTemplateResponse<Accounts> {
     let mut app = custom_mock_app();
@@ -56,13 +53,13 @@ pub fn vending_minter_template(num_tokens: u32) -> MinterTemplateResponse<Accoun
 
 pub fn vending_minter_with_sudo_update_params_template(
     num_tokens: u32,
-) -> MinterTemplateResponse<Accounts> {
+) -> MinterTemplateResponseCodeIds<Accounts> {
     let mut app = custom_mock_app();
     let (creator, buyer) = setup_accounts(&mut app);
     let start_time = Timestamp::from_nanos(GENESIS_MINT_START_TIME);
     let collection_params = mock_collection_params_1(Some(start_time));
     let minter_params = minter_params_token(num_tokens);
-    let code_ids = vending_minter_code_ids(&mut app);
+    let code_ids = base_minter_sg721_collection_code_ids(&mut app);
     let minter_collection_response: Vec<MinterCollectionResponse> = configure_minter(
         &mut app,
         creator.clone(),
@@ -70,11 +67,11 @@ pub fn vending_minter_with_sudo_update_params_template(
         vec![minter_params],
         code_ids.clone(),
     );
-    sudo_update_params(&mut app, &minter_collection_response, code_ids);
-    MinterTemplateResponse {
+    MinterTemplateResponseCodeIds {
         router: app,
         collection_response_vec: minter_collection_response,
         accts: Accounts { creator, buyer },
+        code_ids,
     }
 }
