+++ test-suite/src/vending_minter/tests/happy_unhappy.rs
@@ -1,7 +1,5 @@
 use crate::common_setup::setup_minter::vending_minter::mock_params::mock_create_minter;
-use crate::common_setup::templates::{
-    vending_minter_template, vending_minter_with_sudo_update_params_template,
-};
+use crate::common_setup::templates::vending_minter_template;
 use crate::common_setup::{
     setup_accounts_and_block::coins_for_msg, setup_accounts_and_block::setup_block_time,
 };
@@ -299,45 +297,4 @@ fn unhappy_path() {
     let mint_msg = ExecuteMsg::Mint {};
     let res = router.execute_contract(buyer, minter_addr, &mint_msg, &coins(MINT_PRICE, "uatom"));
     assert!(res.is_err());
-}
-
-#[test]
-fn happy_path_with_params_update() {
-    let vt = vending_minter_with_sudo_update_params_template(2);
-    let (mut router, creator, buyer) = (vt.router, vt.accts.creator, vt.accts.buyer);
-
-    let minter_addr = vt.collection_response_vec[0].minter.clone().unwrap();
-    let collection_addr = vt.collection_response_vec[0].collection.clone().unwrap();
-
-    // Default start time genesis mint time
-    let res: StartTimeResponse = router
-        .wrap()
-        .query_wasm_smart(minter_addr.clone(), &QueryMsg::StartTime {})
-        .unwrap();
-    assert_eq!(
-        res.start_time,
-        Timestamp::from_nanos(GENESIS_MINT_START_TIME).to_string()
-    );
-
-    setup_block_time(&mut router, GENESIS_MINT_START_TIME + 1, None);
-
-    // Fail with incorrect tokens
-    let mint_msg = ExecuteMsg::Mint {};
-    let err = router.execute_contract(
-        buyer.clone(),
-        minter_addr.clone(),
-        &mint_msg,
-        &coins(MINT_PRICE + 100, NATIVE_DENOM),
-    );
-    assert!(err.is_err());
-
-    // Succeeds if funds are sent
-    let mint_msg = ExecuteMsg::Mint {};
-    let res = router.execute_contract(
-        buyer.clone(),
-        minter_addr.clone(),
-        &mint_msg,
-        &coins(MINT_PRICE, NATIVE_DENOM),
-    );
-    assert!(res.is_ok());
 }
